+++ CHANGELOG.md
@@ -36,7 +36,7 @@ which use the proxied NSX-V API of advanced edge gateway for handling NAT rules 
 * Deprecated methods `Client.FindVMByHREF`, `Vdc.FindVMByName`, `Vdc.FindVAppByID`, and `Vdc.FindVAppByName`
 * Added methods `Vm.GetGuestCustomizationSection` and `Vm.SetGuestCustomizationSection`  
 * Deprecated methods `VDC.FindDiskByHREF` and `FindDiskByHREF`
-* Added methods `VDC.GetDiskByHref` `VDC.GetDiskByName` and related `GetDiskById`, `GetDiskByNameOrId`
+* Added methods `VDC.GetDiskByHref` `VDC.GetDisksByName` and related `GetDiskById`
 
 IMPROVEMENTS: